Исправил замечания ревью - Поправил название MovieQuizModelDelegat -> MovieQuizModelDelegate - Полностью перенёс работу с MovieQuizModel в MovieQuizPresenter - Блокирую кнопки и проверяю это в тестах - Вынес моки в отдельные файлы - Исправил название теста testExample -> testGameLogic
