+++ MovieQuiz/Presentation/MovieQuizPresenter/MovieQuizPresenter.swift
@@ -1,4 +1,6 @@
 protocol MovieQuizPresenter {
     var viewController: MovieQuizPresenterDelegate? { get set }
+    func startNewGame()
+    func checkAnswer(_ answer: Answer)
     func updateViewState(to state: GameState)
 }
